rename server.js to appServer
